+++ src/whatsapp/dto/chat.dto.ts
@@ -100,3 +100,9 @@ export class SendPresenceDto extends Metadata {
     delay: number;
   };
 }
+
+export class UpdateMessageDto extends Metadata {
+  number: string;
+  key: proto.IMessageKey;
+  text: string;
+}